+++ src/X/CodeGen/Emitter.java
@@ -354,7 +354,7 @@ public class Emitter implements Visitor {
     public Object visitOperator(Operator ast, Object o) {
         Frame f = (Frame) o;
         switch (ast.spelling) {
-            case "c-", "i-", "f-" -> {
+            case "i8-", "i64-", "f32-" -> {
                 if (ast.parent instanceof BinaryExpr parent) {
                     int numOne = parent.E1.tempIndex;
                     int numTwo = parent.E2.tempIndex;
@@ -384,8 +384,7 @@ public class Emitter implements Visitor {
                     emitN("\t%" + newNum + " = " +  "xor i1 1, " + " %" + numOne);
                 }
             }
-            case "i+", "f+", "c+", "i*", "f*", "c*", "i%", "ii%", "c%", "i/", "f/", "c/", "i==", "f==", "b==", "c==", "i!=", "f!=", "c!=", "b!=",
-                "i<", "f<", "c<", "i<=", "f<=", "c<=", "i>", "f>", "c>", "i>=", "f>=", "c>=" -> {
+            default -> {
                 if (ast.parent instanceof BinaryExpr parent) {
                     int numOne = parent.E1.tempIndex;
                     int numTwo = parent.E2.tempIndex;
@@ -395,44 +394,43 @@ public class Emitter implements Visitor {
                         " %" + numOne + ", %" + numTwo);
                 }
             }
-            default -> {}
         }
         return null;
     }
 
     public String opToCommand(String input) {
         return switch (input)  {
-            case "i+" ->  "add i64";
-            case "c+" ->  "add i8";
-            case "f+" ->  "fadd float";
-            case "i*" -> "mul i64";
-            case "c*" -> "mul i8";
-            case "f*" -> "fmul float";
-            case "i%", "ii%" -> "srem i64";
-            case "c%" -> "srem i8";
-            case "i/" -> "sdiv i64";
-            case "c/" -> "sdiv i8";
-            case "f/" -> "fdiv float";
-            case "i==" -> "icmp eq i64";
-            case "c==" -> "icmp eq i8";
-            case "f==" -> "fcmp oeq float";
+            case "i64+" ->  "add i64";
+            case "i8+" ->  "add i8";
+            case "f32+" ->  "fadd float";
+            case "i64*" -> "mul i64";
+            case "i8*" -> "mul i8";
+            case "f32*" -> "fmul float";
+            case "i64%", "ii%" -> "srem i64";
+            case "i8%" -> "srem i8";
+            case "i64/" -> "sdiv i64";
+            case "i8/" -> "sdiv i8";
+            case "f32/" -> "fdiv float";
+            case "i64==" -> "icmp eq i64";
+            case "i8==" -> "icmp eq i8";
+            case "f32==" -> "fcmp oeq float";
             case "b==" -> "icmp eq i1";
-            case "i!=" -> "icmp ne i64";
-            case "c!=" -> "icmp ne i8";
-            case "f!=" -> "fcmp one float";
+            case "i64!=" -> "icmp ne i64";
+            case "i8!=" -> "icmp ne i8";
+            case "f32!=" -> "fcmp one float";
             case "b!=" -> "icmp ne i1";
-            case "i<=" -> "icmp sle i64";
-            case "c<=" -> "icmp sle i8";
-            case "f<=" -> "fcmp ole float";
-            case "i<" -> "icmp slt i64";
-            case "c<" -> "icmp slt i8";
-            case "f<" -> "fcmp olt float";
-            case "i>" -> "icmp sgt i64";
-            case "c>" -> "icmp sgt i8";
-            case "f>" -> "fcmp ogt float";
-            case "i>=" -> "icmp sge i64";
-            case "c>=" -> "icmp sge i8";
-            case "f>=" -> "fcmp oge float";
+            case "i64<=" -> "icmp sle i64";
+            case "i8<=" -> "icmp sle i8";
+            case "f32<=" -> "fcmp ole float";
+            case "i64<" -> "icmp slt i64";
+            case "i8<" -> "icmp slt i8";
+            case "f32<" -> "fcmp olt float";
+            case "i64>" -> "icmp sgt i64";
+            case "i8>" -> "icmp sgt i8";
+            case "f32>" -> "fcmp ogt float";
+            case "i64>=" -> "icmp sge i64";
+            case "i8>=" -> "icmp sge i8";
+            case "f32>=" -> "fcmp oge float";
             default -> {
                 System.out.println("opToCommand not implemented: " + input);
                 yield "";
@@ -534,7 +532,7 @@ public class Emitter implements Visitor {
         return null;
     }
 
-    public Object visitIntExpr(IntExpr ast, Object o) {
+    public Object visitI64Expr(I64Expr ast, Object o) {
         Frame f = (Frame) o;
         String value = ast.IL.spelling;
         int num = f.getNewIndex();
@@ -543,11 +541,25 @@ public class Emitter implements Visitor {
         return null;
     }
 
-    public Object visitIntType(SignedIntType ast, Object o) {
-        emit(LLVM.INT_TYPE);
+   public Object visitI32Expr(I32Expr ast, Object o) {
+        Frame f = (Frame) o;
+        String value = ast.IL.spelling;
+        int num = f.getNewIndex();
+        emitN("\t%" + num + " = add i32 0, " + value);
+        ast.tempIndex = num;
+        return null;
+    }
+
+    public Object visitI64Type(I64Type ast, Object o) {
+        emit(LLVM.I64_TYPE);
         return null;
     }
     
+    public Object visitI32Type(I32Type ast, Object o) {
+        emit(LLVM.I32_TYPE);
+        return null;
+    }
+
     public Object visitParaList(ParaList ast, Object o) {
         emit("(");
         List list = ast;
@@ -654,7 +666,7 @@ public class Emitter implements Visitor {
             int newIndex = f.getNewIndex();
             emit("\t%" + newIndex + " = ");
             Type T = p.T;
-            if (T.isInt() || T.isBoolean() || T.isChar() || T.isEnum()) {
+            if (T.isI64() || T.isBoolean() || T.isI8() || T.isEnum()) {
                 emit("add ");
                 T.visit(this, o);
                 emit(" 0, ");
@@ -672,7 +684,7 @@ public class Emitter implements Visitor {
         return null;
    }
 
-    public void handleOutInt(CallExpr ast, Object o) {
+    public void handleOutI64(CallExpr ast, Object o) {
         Frame f = (Frame) o;
         ((Args) ast.AL).E.visit(this, o);
         int index = ((Args) ast.AL).E.tempIndex;
@@ -740,8 +752,8 @@ public class Emitter implements Visitor {
             return null;
         }
 
-        if (ast.I.spelling.equals("outInt")) {
-            handleOutInt(ast, o);
+        if (ast.I.spelling.equals("outI64")) {
+            handleOutI64(ast, o);
             return null;
         }
 
@@ -941,8 +953,8 @@ public class Emitter implements Visitor {
         return null;
     }
 
-    public Object visitFloatType(FloatType ast, Object o) {
-        emit(LLVM.FLOAT_TYPE);
+    public Object visitF32Type(F32Type ast, Object o) {
+        emit(LLVM.F32_TYPE);
         return null;
     }
 
@@ -1059,10 +1071,10 @@ public class Emitter implements Visitor {
 
         Type t = ast.type;
         Frame f = (Frame) o;
+        emitN("\t;A");
         ast.index.visit(this, o);
+        emitN("\t;B");
         int index = f.localVarIndex - 1;
-        int newV = f.getNewIndex();
-        emitN("\t%" + newV + " = trunc i64 %" + index + " to i32");
         int newIndex = f.getNewIndex();
         emit("\t%" + newIndex + " = getelementptr ");
         if (!ast.I.decl.isParaDecl()) {
@@ -1094,7 +1106,7 @@ public class Emitter implements Visitor {
             emit(", i32 0");
         }
  
-        emitN(", i32 %" + newV);
+        emitN(", i32 %" + index);
 
         ast.tempIndex = newIndex;
         Type innerT = ((ArrayType) ast.type).t;
@@ -1106,8 +1118,8 @@ public class Emitter implements Visitor {
         return null;
     }
 
-    public Object visitCharType(CharType ast, Object o) {
-        emit(LLVM.CHAR_TYPE);
+    public Object visitI8Type(I8Type ast, Object o) {
+        emit(LLVM.I8_TYPE);
         return null;
     }
 
@@ -1129,20 +1141,52 @@ public class Emitter implements Visitor {
         ast.tempIndex = temp;
         Type from = ast.tFrom, to = ast.tTo;
         emit("\t%" + temp+ " = ");
-        if (from.isInt() && to.isFloat()) {
-            emitN("sitofp i64 %" + numOne + " to float");
+
+        if (from.isI64()) {
+            if (to.isI32()) {
+                emit("trunc i64 %" + numOne + " to i32");
+            } else if (to.isI8()) {
+                emit("trunc i64 %" + numOne + " to i8");
+            } else if (to.isF32()) {
+                emit("sitofp i64 %" + numOne + " to float");
+            }
         }
-        if (from.isChar() && to.isInt()) {
-            emitN("sext i8 %" + numOne + " to i64");
+
+        if (from.isI32()) {
+            if (to.isI64()) {
+                emit("sext i32 %" + numOne + " to i64");
+            } else if (to.isI8()) {
+                emit("trunc i32 %" + numOne + " to i8");
+            } else if (to.isF32()) {
+                emit("sitofp i32 %" + numOne + " to float");
+            }
         }
-        if (from.isInt() && to.isChar()) {
-            emitN("trunc i64 %" + numOne + " to i8");
+
+        if (from.isI8()) {
+            if (to.isI64()) {
+                emit("sext i8 %" + numOne + " to i64");
+            } else if (to.isI32()) {
+                emit("sext i8 %" + numOne + " to i32");
+            } else if (to.isF32()) {
+                emit("sitofp i8 %" + numOne + " to float");
+            }
         }
+
+        if (from.isF32()) {
+            if (to.isI64()) {
+                emit("fptosi float %" + numOne + " to i64");
+            } else if (to.isI32()) {
+                emit("fptosi float %" + numOne + " to i32");
+            } else if (to.isI8()) {
+                emit("fptosi float %" + numOne + " to i8");
+            }
+        }
+
         return null;
     }
 
     public Object visitEnumType(EnumType ast, Object o) {
-        emit(LLVM.INT_TYPE);
+        emit(LLVM.I64_TYPE);
         return null;
     }
 
@@ -1410,12 +1454,10 @@ public class Emitter implements Visitor {
         String structName = "struct." + file + "." + ast.ref.I.spelling;
 
         int arrayIndexNum = -1;
-        int newV = -1;
         if (ast.arrayIndex.isPresent()) {
             ast.arrayIndex.get().visit(this, o);
             arrayIndexNum = f.localVarIndex - 1;
-            newV = f.getNewIndex();
-            emitN("\t%" + newV + " = trunc i64 %" + arrayIndexNum + " to i32");
+          
         }
 
 
@@ -1428,7 +1470,7 @@ public class Emitter implements Visitor {
                 ast.sourceType.visit(this, o);
                 emit(", ");
                 ast.sourceType.visit(this, o);
-                emitN("* %" + localRef + ", i32 0, i32 %" + newV);
+                emitN("* %" + localRef + ", i32 0, i32 %" + arrayIndexNum);
                 localRef = String.valueOf(v);
                 v = f.getNewIndex();
             }
@@ -1457,7 +1499,7 @@ public class Emitter implements Visitor {
                 ast.sourceType.visit(this, o);
                 emit(", ");
                 ast.sourceType.visit(this, o);
-                emitN("* %" + localRef + ", i32 0, i32 %" + newV);
+                emitN("* %" + localRef + ", i32 0, i32 %" + arrayIndexNum);
                 localRef = String.valueOf(v);
                 v = f.getNewIndex();
             }
@@ -1518,11 +1560,11 @@ public class Emitter implements Visitor {
     public void emitBase(Type t, Object o) {
         // TODO: handle other cases
         Expr I;
-        if (t.isInt()) {
-            I = new IntExpr(new IntLiteral("0", dummyPos), dummyPos);
+        if (t.isI64()) {
+            I = new I64Expr(new IntLiteral("0", dummyPos), dummyPos);
         } else if (t.isBoolean()) {
             I = new BooleanExpr(new BooleanLiteral("false", dummyPos), dummyPos);
-        } else if (t.isFloat()) {
+        } else if (t.isF32()) {
             I = new FloatExpr(new FloatLiteral("1.0", dummyPos), dummyPos);
         } else {
             return;
@@ -1598,11 +1640,11 @@ public class Emitter implements Visitor {
         Frame f = (Frame) o;
         int size = -1;
 
-        if (t.isInt() || t.isEnum() || t.isFloat()) {
+        if (t.isI64() || t.isEnum() || t.isF32()) {
             size = 8;
-        } else if (t.isFloat()) {
+        } else if (t.isF32()) {
             size = 4;
-        }else if (t.isChar() || t.isBoolean()) {
+        }else if (t.isI8() || t.isBoolean()) {
             size = 1;
         } else if (t.isPointer()) {
             size = 8;
@@ -1729,4 +1771,7 @@ public class Emitter implements Visitor {
         return null;
     }
 
+    public Object visitIntExpr(IntExpr ast, Object o) {
+        return null;
+    }   
 }