+++ src/X/CodeGen/LLVM.java
@@ -49,9 +49,10 @@ public class LLVM {
         }
     }
 
-    public static final String INT_TYPE = "i64";
-    public static final String CHAR_TYPE = "i8";
-    public static final String FLOAT_TYPE = "float";
+    public static final String I64_TYPE = "i64";
+    public static final String I32_TYPE = "i32";
+    public static final String I8_TYPE = "i8";
+    public static final String F32_TYPE = "float";
     public static final String BOOL_TYPE = "i1";
     public static final String VOID_TYPE = "void";