+++ src/X/Nodes/I32Expr.java
@@ -2,18 +2,18 @@ package X.Nodes;
 
 import X.Lexer.Position;
 
-public class IntExpr extends Expr {
+public class I32Expr extends Expr {
 
     public IntLiteral IL;
 
-    public IntExpr(IntLiteral ilAST, Position pos) {
+    public I32Expr(IntLiteral ilAST, Position pos) {
         super(pos);
         IL = ilAST;
         IL.parent = this;
     }
 
     public Object visit(Visitor v, Object o) {
-        return v.visitIntExpr(this, o);
+        return v.visitI32Expr(this, o);
     }
+    
 }
-